Fixed optional filename bug
Fixed bug where filename was unintentionally being required. It appears that if filename is specified in the add_argument function, it is required even if you set the value to false. I also modified the mysmtp class so that filename was not part of the constructor. If it is specified in __init__ there must be a value passed in when the object is created. I created a filename attribute, but set it to null at object creation. There is a conditional towards that bottom that checks if there was a filename passed in. If there was, it updates the filename attribute which in turn will activate the conditional inside of send_message that actually attaches the file. Specifying a filename upon executing the script will attach the file, if no filename is specified a message will be sent without an attachment.